move blending fxns to blend.rs, add failing tests
