+++ src/colorize.rs
@@ -1,3 +1,4 @@
+use crate::blend;
 use numpy::ndarray::{Array, Array3, ArrayView2};
 
 // Create a (y, x, 3) array with ones
@@ -40,16 +41,16 @@ pub fn apply_colors_and_merge(
                 px_vals.push(color);
                 let color: [u8; 3] = match blending {
                     "max" => {
-                        max_blending(&px_vals)
+                        blend::max_blending(&px_vals)
                     }
                     "sum" => {
-                        sum_blending(&px_vals)
+                        blend::sum_blending(&px_vals)
                     }
                     "min" => {
-                        min_blending(&px_vals)
+                        blend::min_blending(&px_vals)
                     }
                     "mean" => {
-                        mean_blending(&px_vals)
+                        blend::mean_blending(&px_vals)
                     }
                     _ => panic!("received invalid argument for `blending`: {blending}, valid arguments are 'max', 'sum', 'min', and 'mean'")
                 };
@@ -60,68 +61,4 @@ pub fn apply_colors_and_merge(
         }
     }
     rgb
-}
-
-fn max_blending(px_vals: &Vec<[u8; 3]>) -> [u8; 3] {
-    let mut r: u8 = 0;
-    let mut g: u8 = 0;
-    let mut b: u8 = 0;
-    for px_val in px_vals {
-        if px_val[0] > r {
-            r = px_val[0]
-        }
-        if px_val[1] > g {
-            g = px_val[1]
-        }
-        if px_val[2] > b {
-            b = px_val[2]
-        }
-    }
-    [r, g, b]
-}
-
-fn sum_blending(px_vals: &Vec<[u8; 3]>) -> [u8; 3] {
-    let mut r: u8 = 0;
-    let mut g: u8 = 0;
-    let mut b: u8 = 0;
-    for px_val in px_vals {
-        r = r.saturating_add(px_val[0]);
-        g = g.saturating_add(px_val[1]);
-        b = b.saturating_add(px_val[2]);
-    }
-    [r, g, b]
-}
-
-fn min_blending(px_vals: &Vec<[u8; 3]>) -> [u8; 3] {
-    let mut r: u8 = 255;
-    let mut g: u8 = 255;
-    let mut b: u8 = 255;
-    for px_val in px_vals {
-        if px_val[0] < r {
-            r = px_val[0]
-        }
-        if px_val[1] < g {
-            g = px_val[1]
-        }
-        if px_val[2] < b {
-            b = px_val[2]
-        }
-    }
-    [r, g, b]
-}
-
-fn mean_blending(px_vals: &Vec<[u8; 3]>) -> [u8; 3] {
-    let mut r: u16 = 0;
-    let mut g: u16 = 0;
-    let mut b: u16 = 0;
-    for px_val in px_vals {
-        r = r.saturating_add(px_val[0] as u16);
-        g = g.saturating_add(px_val[1] as u16);
-        b = b.saturating_add(px_val[2] as u16);
-    }
-    let n_channels: u16 = px_vals.len() as u16;
-    let r: u8 = (r / n_channels) as u8;
-    let g: u8 = (g / n_channels) as u8;
-    let b: u8 = (b / n_channels) as u8;
-    [r, g, b]
 }
